requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-dotenv==1.0.0
motor==3.3.1
groq==0.4.1
sentence-transformers==2.2.2
langgraph==0.0.24
pydantic==2.4.2
wikipedia==1.4.0
pymongo[srv]==4.6.0
python-multipart==0.0.6
websockets==11.0.3
httpx==0.24.1
pytest==7.4.3
pytest-asyncio==0.21.1
langchain==0.0.339
langchain-groq==0.0.1
numpy==1.24.3
tiktoken==0.5.1
setuptools==68.2.2
matplotlib==3.7.2
beautifulsoup4==4.12.2
requests==2.31.0 